rm unnecessary cast

+++ src/packet/data_types/var_int.rs
@@ -26,7 +26,7 @@ impl TryFrom<VarInt> for isize {
                 if i > 4 {
                     return Err(VarIntParseError::WrongSize);
                 }
-                return Ok(result as isize);
+                return Ok(result);
             }
         }
